requirements_server_ndd.txt
opencv-python
numpy
tqdm
datetime
tensorflow
keras
scipy
pillow
argparse
sklearn
tensorflow